Fix decoding of TCAP messages
Recent versions of asn1c fail while decoding the DialoguePortion of a TCAP
message - probably because Tag APPLICATION 11 is sent as constructed tag and
the WS adaptation expects a primitive one. Not sure if this ever worked
with an older asn1c version, maybe it ignored this discrepancy.

The described behavior also happens with mouse07410/asn1c 2c06555 used
in the next patch to regenerate the code.

In any case the correct encoding of the DialoguePortion is as EXTERNAL.
Support for proper EXTERNAL asn1c decoding is missing in all versions of asn1c
except https://github.com/mouse07410/asn1c (See PR #3/#5) so use that
when regenerating.

Related: SYS#7486
Change-Id: I4a27a52b2be70b1fc50e0a3d5441ded4c151fe65